Godeps: update docker2aci to v0.9.1.
This pulls in the latest docker2aci.

Fixes #2248.
