+++ Godeps/_workspace/src/github.com/appc/docker2aci/lib/conversion_store.go
@@ -1,3 +1,17 @@
+// Copyright 2015 The appc Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package docker2aci
 
 import (
@@ -23,18 +37,18 @@ type aciInfo struct {
 	ImageManifest *schema.ImageManifest
 }
 
-// ConversionStore is an simple implementation of the acirenderer.ACIRegistry
+// conversionStore is an simple implementation of the acirenderer.ACIRegistry
 // interface. It stores the Docker layers converted to ACI so we can take
 // advantage of acirenderer to generate a squashed ACI Image.
-type ConversionStore struct {
+type conversionStore struct {
 	acis map[string]*aciInfo
 }
 
-func NewConversionStore() *ConversionStore {
-	return &ConversionStore{acis: make(map[string]*aciInfo)}
+func newConversionStore() *conversionStore {
+	return &conversionStore{acis: make(map[string]*aciInfo)}
 }
 
-func (ms *ConversionStore) WriteACI(path string) (string, error) {
+func (ms *conversionStore) WriteACI(path string) (string, error) {
 	f, err := os.Open(path)
 	if err != nil {
 		return "", err
@@ -65,7 +79,7 @@ func (ms *ConversionStore) WriteACI(path string) (string, error) {
 	return key, nil
 }
 
-func (ms *ConversionStore) GetImageManifest(key string) (*schema.ImageManifest, error) {
+func (ms *conversionStore) GetImageManifest(key string) (*schema.ImageManifest, error) {
 	aci, ok := ms.acis[key]
 	if !ok {
 		return nil, fmt.Errorf("aci with key: %s not found", key)
@@ -73,7 +87,7 @@ func (ms *ConversionStore) GetImageManifest(key string) (*schema.ImageManifest, 
 	return aci.ImageManifest, nil
 }
 
-func (ms *ConversionStore) GetACI(name types.ACIdentifier, labels types.Labels) (string, error) {
+func (ms *conversionStore) GetACI(name types.ACIdentifier, labels types.Labels) (string, error) {
 	for _, aci := range ms.acis {
 		// we implement this function to comply with the interface so don't
 		// bother implementing a proper label check
@@ -84,7 +98,7 @@ func (ms *ConversionStore) GetACI(name types.ACIdentifier, labels types.Labels) 
 	return "", fmt.Errorf("aci not found")
 }
 
-func (ms *ConversionStore) ReadStream(key string) (io.ReadCloser, error) {
+func (ms *conversionStore) ReadStream(key string) (io.ReadCloser, error) {
 	img, ok := ms.acis[key]
 	if !ok {
 		return nil, fmt.Errorf("stream for key: %s not found", key)
@@ -102,11 +116,11 @@ func (ms *ConversionStore) ReadStream(key string) (io.ReadCloser, error) {
 	return tr, nil
 }
 
-func (ms *ConversionStore) ResolveKey(key string) (string, error) {
+func (ms *conversionStore) ResolveKey(key string) (string, error) {
 	return key, nil
 }
 
-func (ms *ConversionStore) HashToKey(h hash.Hash) string {
+func (ms *conversionStore) HashToKey(h hash.Hash) string {
 	s := h.Sum(nil)
 	return fmt.Sprintf("%s%x", hashPrefix, s)
 }
