+++ Godeps/_workspace/src/github.com/appc/docker2aci/lib/internal/backend/repository/repository.go
@@ -1,4 +1,4 @@
-// Copyright 2015 CoreOS, Inc.
+// Copyright 2015 The appc Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -11,6 +11,12 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
+// Package repository is an implementation of Docker2ACIBackend for Docker
+// remote registries.
+//
+// Note: this package is an implementation detail and shouldn't be used outside
+// of docker2aci.
 package repository
 
 import (
@@ -20,7 +26,8 @@ import (
 	"path"
 
 	"github.com/appc/docker2aci/lib/common"
-	"github.com/appc/docker2aci/lib/types"
+	"github.com/appc/docker2aci/lib/internal/docker"
+	"github.com/appc/docker2aci/lib/internal/types"
 	"github.com/appc/spec/schema"
 )
 
@@ -54,7 +61,7 @@ func NewRepositoryBackend(username string, password string, insecure bool) *Repo
 }
 
 func (rb *RepositoryBackend) GetImageInfo(url string) ([]string, *types.ParsedDockerURL, error) {
-	dockerURL, err := common.ParseDockerURL(url)
+	dockerURL, err := docker.ParseDockerURL(url)
 	if err != nil {
 		return nil, nil, err
 	}
