+++ Godeps/_workspace/src/github.com/appc/docker2aci/lib/internal/backend/repository/repository2.go
@@ -1,4 +1,4 @@
-// Copyright 2015 CoreOS, Inc.
+// Copyright 2015 The appc Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -11,6 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
 package repository
 
 import (
@@ -29,8 +30,9 @@ import (
 	"time"
 
 	"github.com/appc/docker2aci/lib/common"
-	"github.com/appc/docker2aci/lib/types"
-	"github.com/appc/docker2aci/lib/util"
+	"github.com/appc/docker2aci/lib/internal"
+	"github.com/appc/docker2aci/lib/internal/types"
+	"github.com/appc/docker2aci/pkg/log"
 	"github.com/appc/spec/schema"
 	"github.com/coreos/ioprogress"
 )
@@ -93,8 +95,8 @@ func (rb *RepositoryBackend) buildACIV2(layerNumber int, layerID string, dockerU
 	}
 	defer layerFile.Close()
 
-	util.Debug("Generating layer ACI...")
-	aciPath, aciManifest, err := common.GenerateACI(layerNumber, layerData, dockerURL, outputDir, layerFile, curPwl, compression)
+	log.Debug("Generating layer ACI...")
+	aciPath, aciManifest, err := internal.GenerateACI(layerNumber, layerData, dockerURL, outputDir, layerFile, curPwl, compression)
 	if err != nil {
 		return "", nil, fmt.Errorf("error generating ACI: %v", err)
 	}
@@ -358,7 +360,8 @@ func (rb *RepositoryBackend) makeRequest(req *http.Request, repo string) (*http.
 	}
 	// The scope can be empty if we're not getting a token for a specific repo
 	if scope == "" && repo != "" {
-		return nil, fmt.Errorf("missing scope in bearer auth challenge")
+		// If the scope is empty and it shouldn't be, we can infer it based on the repo
+		scope = fmt.Sprintf("repository:%s:pull", repo)
 	}
 
 	authReq, err := http.NewRequest("GET", realm, nil)
