+++ Godeps/_workspace/src/github.com/appc/docker2aci/lib/internal/types/types.go
@@ -1,0 +1,26 @@
+// Copyright 2015 The appc Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Package types defines Docker image, URL and configuration types.
+//
+// Note: this package is an implementation detail and shouldn't be used outside
+// of docker2aci.
+package types
+
+// ParsedDockerURL represents a parsed Docker URL.
+type ParsedDockerURL struct {
+	IndexURL  string
+	ImageName string
+	Tag       string
+}
