--- Godeps/_workspace/src/google.golang.org/grpc/grpclog/glogger/glogger.go
@@ -1,72 +1,0 @@
-/*
- *
- * Copyright 2015, Google Inc.
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are
- * met:
- *
- *     * Redistributions of source code must retain the above copyright
- * notice, this list of conditions and the following disclaimer.
- *     * Redistributions in binary form must reproduce the above
- * copyright notice, this list of conditions and the following disclaimer
- * in the documentation and/or other materials provided with the
- * distribution.
- *     * Neither the name of Google Inc. nor the names of its
- * contributors may be used to endorse or promote products derived from
- * this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- */
-
-/*
-Package glogger defines glog-based logging for grpc.
-*/
-package glogger
-
-import (
-	"github.com/golang/glog"
-	"google.golang.org/grpc/grpclog"
-)
-
-func init() {
-	grpclog.SetLogger(&glogger{})
-}
-
-type glogger struct{}
-
-func (g *glogger) Fatal(args ...interface{}) {
-	glog.Fatal(args...)
-}
-
-func (g *glogger) Fatalf(format string, args ...interface{}) {
-	glog.Fatalf(format, args...)
-}
-
-func (g *glogger) Fatalln(args ...interface{}) {
-	glog.Fatalln(args...)
-}
-
-func (g *glogger) Print(args ...interface{}) {
-	glog.Info(args...)
-}
-
-func (g *glogger) Printf(format string, args ...interface{}) {
-	glog.Infof(format, args...)
-}
-
-func (g *glogger) Println(args ...interface{}) {
-	glog.Infoln(args...)
-}
